+++ weblate/checks/tests/test_placeholders.py
@@ -19,12 +19,13 @@
 
 """Tests for placeholder quality checks."""
 
-
 from weblate.checks.flags import Flags
 from weblate.checks.models import Check
 from weblate.checks.placeholders import PlaceholderCheck, RegexCheck
 from weblate.checks.tests.test_checks import CheckTestCase, MockUnit
+from weblate.lang.models import Language
 from weblate.trans.models import Unit
+from weblate.trans.tests.test_views import FixtureTestCase
 
 
 class PlaceholdersTest(CheckTestCase):
@@ -123,6 +124,43 @@ class PlaceholdersTest(CheckTestCase):
                     self.default_lang,
                     "Hello %WORLD%",
                 ),
+            )
+        )
+
+
+class PluralPlaceholdersTest(FixtureTestCase):
+    def test_plural(self):
+        check = PlaceholderCheck()
+        lang = "cs"
+        unit = MockUnit(
+            None,
+            'placeholders:r"%[0-9]"',
+            lang,
+            "1 apple",
+        )
+        unit.translation.language = Language.objects.get(code=lang)
+        unit.translation.plural = unit.translation.language.plural
+        self.assertFalse(
+            check.check_target(
+                ["1 apple", "%1 apples"],
+                ["1 jablko", "%1 jablka", "%1 jablek"],
+                unit,
+            )
+        )
+        unit.check_cache = {}
+        self.assertTrue(
+            check.check_target(
+                ["1 apple", "%1 apples"],
+                ["1 jablko", "1 jablka", "%1 jablek"],
+                unit,
+            )
+        )
+        unit.check_cache = {}
+        self.assertTrue(
+            check.check_target(
+                ["%1 apple", "%1 apples"],
+                ["1 jablko", "1 jablka", "%1 jablek"],
+                unit,
             )
         )