+++ docs/changes.rst
@@ -14,6 +14,7 @@ Not yet released.
 * Added support for adding custom markup to HTML head.
 * Fixed generation of MO files in the add-on to include only translated files.
 * Fixed rendering of regular expression flags.
+* Improved placeholders check behavior with plurals.
 
 `All changes in detail <https://github.com/WeblateOrg/weblate/milestone/88?closed=1>`__.